# backend/openapi.yaml
openapi: 3.0.0
info:
  title: GameHub API
  version: 1.0.0
  description: API documentation for GameHub Backend
servers:
  - url: https://gamehub-kzzk.onrender.com
    description: Production Server
  - url: https://israel-ramose-premeditatingly.ngrok-free.dev
    description: Ngrok Tunnel
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API Key for general access (x-api-key header)
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT Access Token in cookie
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: JWT Refresh Token in cookie
security:
  - ApiKeyAuth: []
    cookieAuth: []
paths:
...
  /api/friends/request:
    post:
      summary: Send a friend request
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                receiverId:
                  type: string
                  example: friend_id_123
      responses:
        '200':
          description: Request sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Friend request sent
  /api/friends/accept/{id}:
    post:
      summary: Accept a friend request
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: request_id_123
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Friend request accepted
  /api/friends/remove/{id}:
    delete:
      summary: Remove a friend or reject request
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: friend_id_123
      responses:
        '200':
          description: Friend removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Friend removed
  /api/friends/search/{id}:
    get:
      summary: Check friendship status with a user
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: user_to_check
      responses:
        '200':
          description: Friendship status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: friend
  /api/friends/list:
    post:
      summary: Get list of friends (POST variant)
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of friends
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: friend_id_123
                    username:
                      type: string
                      example: friend_name
  /api/friends/requests:
    get:
      summary: Get pending friend requests
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: Pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: request_id
                    sender:
                      type: object
                      properties:
                        username:
                          type: string
                          example: requester_name
  /api/friends/suggestions:
    get:
      summary: Get friend suggestions
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: Suggestions list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                      example: suggested_user
  /api/friends/find:
    get:
      summary: Find user by ID or specific criteria
      tags:
        - Friends
      security:
        - cookieAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
...